Added remote libsql support

+++ lib/pages/todos/todoeditbar/delete_selected_todos_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../models/drift.dart';
+import '../../../models/db_selector.dart';
 import '../../../models/local_state.dart';
 import '../confirmation_dialog.dart';
 
@@ -15,8 +15,11 @@ class DeleteSelectedTodosButton extends StatelessWidget {
       context: context,
       builder: (context) => ConfirmationDialog(),
     );
-    if (shouldDelete != null && shouldDelete) {
-      await MyDatabase.instance.managers.todo
+    if (shouldDelete != null && shouldDelete && context.mounted) {
+      await Provider.of<DbSelector>(context, listen: false)
+          .database
+          .managers
+          .todo
           .filter((f) => f.id.isIn(localStateNotifier.selectedTodoIds))
           .delete();
       localStateNotifier.setSelectedTodoIds([]);
